Fix and unify asset collection for addons.
