+++ website/templates/project/files.mako
@@ -33,7 +33,7 @@ ${parent.stylesheets()}
 <%def name="javascript_bottom()">
 ${parent.javascript_bottom()}
 % for script in tree_js:
-<script type="text/javascript" src="${script | webpack_asset}"></script>
+<script type="text/javascript" src="${script}"></script>
 % endfor
 <script src=${"/static/public/js/files-page.js" | webpack_asset}></script>
 </%def>
