+++ website/templates/project/project.mako
@@ -151,7 +151,7 @@
 ${parent.javascript_bottom()}
 
 % for script in tree_js:
-<script type="text/javascript" src="${script | webpack_asset}"></script>
+<script type="text/javascript" src="${script}"></script>
 % endfor
 
 <script type="text/javascript">
@@ -174,7 +174,7 @@ ${parent.javascript_bottom()}
 <script src="${"/static/public/js/project-dashboard.js" | webpack_asset}"></script>
 
 % for asset in addon_widget_js:
-<script src="${asset | webpack_asset}"></script>
+<script src="${asset}"></script>
 % endfor
 
 </%def>
