+++ website/util/rubeus.py
@@ -11,7 +11,7 @@ import hurry
 from modularodm import Q
 
 from framework.auth.decorators import Auth
-from website import settings
+from website.util.webpack import webpack_asset
 from website.settings import (
     ALL_MY_PROJECTS_ID, ALL_MY_REGISTRATIONS_ID, ALL_MY_PROJECTS_NAME,
     ALL_MY_REGISTRATIONS_NAME
@@ -490,17 +490,11 @@ def collect_addon_js(node, visited=None, filename='files.js', config_entry='file
         # JS modules configured in each addon's __init__ file
         js = js.union(addon.config.include_js.get(config_entry, []))
         # Webpack bundle
-        file_path = os.path.join('static',
-                                 'public',
-                                 'js',
-                                 addon.config.short_name,
-                                 filename)
-        js_file = os.path.join(
-            settings.BASE_PATH, file_path
-        )
-        if os.path.exists(js_file):
-            js_path = os.path.join('/', file_path)
+        try:
+            js_path = webpack_asset(os.path.join(addon.config.short_name, filename))
             js.add(js_path)
+        except KeyError:
+            pass
     for each in node.nodes:
         if each._id not in visited:
             visited.append(each._id)
@@ -511,8 +505,8 @@ def collect_addon_js(node, visited=None, filename='files.js', config_entry='file
 def collect_addon_css(node, visited=None):
     """Collect CSS includes for all addons-ons implementing Hgrid views.
 
-    :return list: List of CSS include paths
-
+    :return: List of CSS include paths
+    :rtype: list
     """
     visited = visited or []
     visited.append(node._id)
